+++ core/designsystem/src/main/java/com/inseoul/designsystem/theme/Type.kt
@@ -1,0 +1,77 @@
+package com.inseoul.onetwothree.ui.theme
+
+import androidx.compose.material.Typography
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+import com.inseoul.designsystem.R
+
+val pretendard = FontFamily(
+    Font(R.font.pretendard_bold, FontWeight.Bold, FontStyle.Normal),
+    Font(R.font.pretendard_medium, FontWeight.Medium, FontStyle.Normal),
+    Font(R.font.pretendard_regular, FontWeight.Normal, FontStyle.Normal),
+    Font(R.font.pretendard_light, FontWeight.Light, FontStyle.Normal),
+    Font(R.font.pretendard_thin, FontWeight.Thin, FontStyle.Normal),
+    Font(R.font.pretendard_extrabold, FontWeight.ExtraBold, FontStyle.Normal),
+    Font(R.font.pretendard_extralight, FontWeight.ExtraLight, FontStyle.Normal),
+    Font(R.font.pretendard_semibold, FontWeight.SemiBold, FontStyle.Normal),
+    Font(R.font.pretendard_black, FontWeight.Black, FontStyle.Normal)
+)
+
+// Set of Material typography styles to start with
+val Typography = Typography(
+
+    h1 = TextStyle(
+        fontFamily = pretendard,
+        fontWeight = FontWeight.Bold,
+        fontSize = 24.sp
+    ),
+
+    h2 = TextStyle(
+        fontFamily = pretendard,
+        fontWeight = FontWeight.Bold,
+        fontSize = 20.sp
+    ),
+
+    subtitle1 = TextStyle(
+        fontFamily = pretendard,
+        fontWeight = FontWeight.Bold,
+        fontSize = 16.sp
+    ),
+
+    body1 = TextStyle(
+        fontFamily = pretendard,
+        fontWeight = FontWeight.Normal,
+        fontSize = 18.sp
+    ),
+
+    body2 = TextStyle(
+        fontFamily = pretendard,
+        fontWeight = FontWeight.Normal,
+        fontSize = 14.sp
+    ) ,
+
+    button = TextStyle(
+        fontFamily = pretendard,
+        fontWeight = FontWeight.Bold,
+        fontSize = 16.sp
+    )
+
+
+
+    /* Other default text styles to override
+    button = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.W500,
+        fontSize = 14.sp
+    ),
+    caption = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 12.sp
+    )
+    */
+)
